refactoring

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { TaskRepository } from './task-repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status-enum';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,8 +15,11 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ){}
     
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    async getTasks(
+        filterDto: GetTasksFilterDto,
+        user: User
+        ): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
     // getAllTasks(): Task[] {
@@ -40,8 +44,11 @@ export class TasksService {
     //     return tasks;
     // }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(
+        id: number,
+        user: User,
+        ): Promise<Task> {
+        const found = await this.taskRepository.findOne({where : {id, userId: user.id}});
 
         if (!found){
             throw new NotFoundException(`Task with ID "${id}" not found `);
@@ -51,13 +58,19 @@ export class TasksService {
 
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task>{
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(
+        createTaskDto: CreateTaskDto,
+        user: User,
+        ): Promise<Task>{
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(
+        id: number,
+        user: User
+        ): Promise<void> {
+        const result = await this.taskRepository.delete({id, userId: user.id});
 
         if (result.affected == 0) {
             throw new NotFoundException(`Task with ID "${id}" not found `);
@@ -65,16 +78,16 @@ export class TasksService {
     }
 
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number,
+        status: TaskStatus,
+        user: User
+         ): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
 
         return task;
     }
-    // updateTaskStatus(id: string, status: TaskStatus): Task {
-    //     const task = this.getTaskById(id);
-    //     task.status = status;
-    //     return task
-    // }
+
 }
